Initial migration to libobs

+++ main.py
@@ -1,25 +1,63 @@
-import asyncio
-import logging
-import pathlib
 import os
-import subprocess
+from subprocess import Popen, PIPE, STDOUT
+import asyncio
+import re
+import decky_plugin
 
-PARENT_DIR = str(pathlib.Path(__file__).parent.resolve())
-
-logging.basicConfig(
-    format = '[deckystream] %(asctime)s %(levelname)s %(message)s')
-
-os.environ['HOME'] = "/home/deck"
-os.environ['XDG_RUNTIME_DIR'] = "/run/user/1000"
-os.environ['LD_LIBRARY_PATH'] = PARENT_DIR + "/bin/lib"
-os.environ['GST_PLUGIN_PATH'] = PARENT_DIR + "/bin/lib/gstreamer"
-
- 
 class Plugin:
+
+    #hack to find xauth file (todo: learn about xauth)
+    def find_uuid_file(directory):
+        pattern = re.compile(
+            r'^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$', re.I
+        )
+
+        for root, dirs, files in os.walk(directory):
+            for basename in files:
+                if pattern.match(basename):
+                    filename = os.path.join(root, basename)
+                    return filename
+
+    def log_subprocess_output(pipe):
+        for line in iter(pipe.readline, b''): # b'\n'-separated lines
+            decky_plugin.logger.info('.NET: %r', line)
+
+    backend_proc = None
+    # Asyncio-compatible long-running code, executed in a task when the plugin is loaded
     async def _main(self):
-        self.backend_proc = subprocess.Popen([PARENT_DIR + "/bin/deckystream"])
+        decky_plugin.logger.info("decky-obs starting!")
+
+        env_proc = dict(os.environ)
+        if "LD_LIBRARY_PATH" in env_proc:
+            env_proc["LD_LIBRARY_PATH"] += ":"+decky_plugin.DECKY_PLUGIN_DIR+"/bin"
+        else:
+            env_proc["LD_LIBRARY_PATH"] = ":"+decky_plugin.DECKY_PLUGIN_DIR+"/bin"
+
+        #env_proc["XAUTHORITY"] = self.find_uuid_file('/run/user/1000/')
+        env_proc["DISPLAY"] = ":0"
+
+        self.backend_proc = Popen(
+            [decky_plugin.DECKY_PLUGIN_DIR + "/bin/obs_recorder"],
+            env = env_proc, stdout=PIPE, stderr=STDOUT)
+        with self.backend_proc.stdout:
+            self.log_subprocess_output(self.backend_proc.stdout)
+
         while True:
             await asyncio.sleep(1)
 
+    # Function called first during the unload process, utilize this to handle your plugin being removed
     async def _unload(self):
-        self.backend_proc.kill()
+        decky_plugin.logger.info("decky-obs closing!")
+        if self.backend_proc is not None:
+            self.backend_proc.terminate()
+            try:
+                self.backend_proc.wait(timeout=5) # 5 seconds timeout
+            except subprocess.TimeoutExpired:
+                self.backend_proc.kill()
+            self.backend_proc = None
+
+        pass
+
+    # Migrations that should be performed before entering `_main()`.
+    # async def _migration(self):
+    #     decky_plugin.logger.info("Migrating")
